etcd_service_publisher/requirements.txt
python-etcd==0.3.2
netifaces==0.10.4
